🚚[Rename] user, bookClub, review, Info 공통 type으로 분리 #253
