+++ src/app/profile/[userId]/page.tsx
@@ -1,5 +1,5 @@
-import { mockUser } from '@/features/profile/constants/mock';
 import { ProfilePage } from '@/features/profile/container';
+import { mockUser } from '@/mocks/mockDatas';
 import React from 'react';
 
 export default function Profile() {
