+++ src/features/profile/components/info/Info.stories.tsx
@@ -1,19 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import Info from './Info';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { User } from '@/types/user';
+import { mockUser } from '@/mocks/mockDatas';
 
 const queryClient = new QueryClient();
-
-const mockUser: User = {
-  id: 123,
-  nickname: 'test',
-  email: 'john.doe@example.com',
-  name: 'John Doe',
-  description: 'Example Company',
-  createdAt: new Date('2024-01-01T00:00:00Z'),
-  updatedAt: new Date('2024-01-10T00:00:00Z'),
-};
 
 const meta = {
   title: 'Components/Info',
